api/requirements.txt
uvicorn~=0.30.6
fastapi~=0.114.0
numpy~=1.26.4
tritonclient~=2.49.0
tritonclient[grpc]~=2.49.0
faiss-cpu==1.8.0
pandas==2.2.3
pydantic==2.9.1
requests==2.32.3
onnxruntime==1.19.2
onnx==1.16.2
